feat: add checkbox component
