+++ src/components/card/card.module.css
@@ -1,6 +1,6 @@
 .characterCard {
   width: 300px;
-  height: 380px;
+  height: 400px;
   padding: 20px;
   text-align: center;
   position: relative;
