+++ src/components/checkbox/checkbox.module.css
@@ -1,0 +1,78 @@
+.checkboxWrapper:hover .check {
+  stroke-dashoffset: 0;
+}
+
+.checkboxWrapper {
+  position: relative;
+  display: inline-block;
+  width: 40px;
+  height: 40px;
+}
+.checkboxWrapper .background {
+  fill: #ccc;
+  transition: ease all 0.6s;
+  -webkit-transition: ease all 0.6s;
+}
+
+.darkThemeCheckboxWrapper .background {
+  fill: white;
+}
+
+.checkboxWrapper .stroke {
+  fill: none;
+  stroke: #fff;
+  stroke-miterlimit: 10;
+  stroke-width: 2px;
+  stroke-dashoffset: 100;
+  stroke-dasharray: 100;
+  transition: ease all 0.6s;
+  -webkit-transition: ease all 0.6s;
+}
+
+.darkThemeCheckboxWrapper .stroke {
+  stroke: #000;
+}
+
+.checkboxWrapper .check {
+  fill: none;
+  stroke: #fff;
+  stroke-linecap: round;
+  stroke-linejoin: round;
+  stroke-width: 2px;
+  stroke-dashoffset: 22;
+  stroke-dasharray: 22;
+  transition: ease all 0.6s;
+  -webkit-transition: ease all 0.6s;
+}
+
+.darkThemeCheckboxWrapper .check {
+  stroke: black;
+}
+
+.checkboxWrapper input[type="checkbox"] {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  left: 0;
+  top: 0;
+  margin: 0;
+  opacity: 0;
+  -appearance: none;
+}
+.checkboxWrapper input[type="checkbox"]:hover {
+  cursor: pointer;
+}
+.checkboxWrapper input[type="checkbox"]:checked + svg .background {
+  fill: black;
+}
+
+.darkThemeCheckboxWrapper input[type="checkbox"]:checked + svg .background {
+  fill: white;
+}
+
+.checkboxWrapper input[type="checkbox"]:checked + svg .stroke {
+  stroke-dashoffset: 0;
+}
+.checkboxWrapper input[type="checkbox"]:checked + svg .check {
+  stroke-dashoffset: 0;
+}
